Add min max curves to the perbasequalityscores plot Add line that shows the reduction in the number of sequeces per length Add quality encodings to  ylabels Add variables for lowest and highest quantiles

+++ uk/ac/babraham/FastQC/Utilities/QualityCount.java
@@ -73,7 +73,7 @@ public class QualityCount {
 		return ((double)total)/count;
 	}
 	
-	public double getPercentile (int offset, int percentile) {
+	public double getPercentile (int offset, double percentile) {
 
 		long total = totalCounts;
